@@ -1,26 +1,30 @@
 import {
-  HashRouter as Router,
-  Route
+  BrowserRouter as Router,
+  Routes,
+  Route,
 } from "react-router-dom";
-
 
 import './App.css';
 import Header from './components/Header'
 import NotesListPage from './pages/NotesListPage'
 import NotePage from './pages/NotePage'
 
+
 function App() {
-  return (
-    <Router>
-      <div className="container dark">
-        <div className="app">
-          <Header />
-          <Route path="/" exact component={NotesListPage} />
-          <Route path="/note/:id" component={NotePage} />
+    return (
+        <div className="container dark">
+            <div className="app">
+                <Header />
+                <Router>
+                    <Routes>
+                        <Route path="/" exact element={ <NotesListPage /> } /> 
+                        <Route path="/note/:id" element={ <NotePage /> } />
+                    </Routes>
+                </Router>
+            </div>
         </div>
-      </div>
-    </Router>
-  );
-}
+    );
+};
+
 
 export default App;
