@@ -1,15 +1,20 @@
-from django.http import response
+from django.http import JsonResponse, response
+from django.middleware.csrf import get_token
 from django.shortcuts import render
+from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework.serializers import Serializer
+
 from .models import Note
 from .serializers import NoteSerializer
 from api import serializers
 from .utils import updateNote, getNoteDetail, deleteNote, getNotesList, createNote
+
 # Create your views here.
 
 
+# This is a function based view that has a rest framework decorator
 @api_view(['GET'])
 def getRoutes(request):
 
@@ -48,15 +53,8 @@ def getRoutes(request):
     return Response(routes)
 
 
-# /notes GET
-# /notes POST
-# /notes/<id> GET
-# /notes/<id> PUT
-# /notes/<id> DELETE
-
 @api_view(['GET', 'POST'])
 def getNotes(request):
-
     if request.method == 'GET':
         return getNotesList(request)
 
@@ -66,7 +64,7 @@ def getNotes(request):
 
 @api_view(['GET', 'PUT', 'DELETE'])
 def getNote(request, pk):
-
+    # Primary key is passed as a parameter
     if request.method == 'GET':
         return getNoteDetail(request, pk)
 
@@ -77,30 +75,39 @@ def getNote(request, pk):
         return deleteNote(request, pk)
 
 
-# @api_view(['POST'])
-# def createNote(request):
-#     data = request.data
-#     note = Note.objects.create(
-#         body=data['body']
-#     )
-#     serializer = NoteSerializer(note, many=False)
-#     return Response(serializer.data)
+@api_view(['POST'])
+def createNote(request):
+    data = request.data
+    note = Note.objects.create(body=data['body'])
+    serializer = NoteSerializer(note, many=False)
+    return Response(serializer.data)
 
 
-# @api_view(['PUT'])
-# def updateNote(request, pk):
-#     data = request.data
-#     note = Note.objects.get(id=pk)
-#     serializer = NoteSerializer(instance=note, data=data)
+# @csrf_exempt
+@api_view(['PUT'])
+def updateNote(request, pk):
+    data = request.data
+    note = Note.objects.get(id=pk)
+    serializer = NoteSerializer(instance=note, data=data)
 
-#     if serializer.is_valid():
-#         serializer.save()
+    if serializer.is_valid():
+        serializer.save()
 
-#     return Response(serializer.data)
+    return Response(serializer.data)
+    # return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-# @api_view(['DELETE'])
-# def deleteNote(request, pk):
-#     note = Note.objects.get(id=pk)
-#     note.delete()
-#     return Response('Note was deleted!')
+@api_view(['DELETE'])
+def deleteNote(request, pk):
+    note = Note.objects.get(id=pk)
+    note.delete()
+    return Response('Note was deleted!')
+
+
+# Set up CSRF token as per this website:
+# https://fractalideas.com/blog/making-react-and-django-play-well-together-single-page-app-model/
+def csrf(request):
+    return JsonResponse({'csrfToken': get_token(request)})
+
+def ping(request):
+    return JsonResponse({'result': 'OK'})
