@@ -3,12 +3,13 @@ import { Link } from 'react-router-dom'
 import { ReactComponent as AddIcon } from '../assets/add.svg'
 
 
+
 const AddButton = () => {
-    return (
-        <Link to="/note/new" className="floating-button">
-            <AddIcon />
-        </Link>
-    )
+	return (
+		<Link to='/note/new' className="floating-button">
+			<AddIcon />
+		</Link>
+	)
 }
 
 export default AddButton
