@@ -1,12 +1,13 @@
 import React from 'react'
 
-const Header = () => {
 
-    return (
-        <div className="app-header">
-            <h1>Note List</h1>
-        </div>
-    )
+const Header = () => {
+	return (
+		<div className="app-header">
+			<h1>Note List</h1>
+		</div>
+	)
 }
+
 
 export default Header
